Move Mock classes into separate package
