+++ src/test/java/http_server/ClientTest.java
@@ -3,6 +3,7 @@
  */
 package http_server;
 
+import Mocks.MockWriter;
 import org.junit.Test;
 
 import java.io.BufferedReader;
